+++ lib/routemaster/middleware/cache.rb
@@ -7,7 +7,7 @@ require 'routemaster/event_index'
 module Routemaster
   module Middleware
     class Cache
-      def initialize(app, cache:nil, client:nil, queue:nil)
+      def initialize(app, cache:nil, client:nil, queue:nil, **_)
         @app    = app
         @cache  = cache || Routemaster::Cache.new
         @client = client || Routemaster::Jobs::Client.new
@@ -16,7 +16,6 @@ module Routemaster
 
       def call(env)
         env.fetch('routemaster.dirty', []).each do |url|
-          @cache.invalidate(url)
           @client.enqueue(@queue, Routemaster::Jobs::CacheAndSweep, url)
         end
         @app.call(env)