+++ spec/routemaster/drain/mapping_spec.rb
@@ -3,6 +3,7 @@ require 'spec/support/rack_test'
 require 'spec/support/uses_redis'
 require 'spec/support/uses_dotenv'
 require 'spec/support/events'
+require 'spec/support/siphon'
 require 'routemaster/drain/mapping'
 require 'json'
 
@@ -10,17 +11,20 @@ describe Routemaster::Drain::Mapping do
   uses_dotenv
   uses_redis
 
-  let(:app) { described_class.new }
+  let(:app) { described_class.new(options) }
+  let(:options) { {} }
   let(:listener) { double 'listener' }
-  
+
   before { app.subscribe(listener, prefix: true) }
 
   let(:path)    { '/' }
-  let(:payload) { [1,2,3,1].map { |idx| make_event(idx) }.to_json }
+  let(:payload) { [1,2,3,1].map { |idx| make_event(idx) } }
   let(:environment) {{ 'CONTENT_TYPE' => 'application/json' }}
-  let(:perform) { post path, payload, environment }
+  let(:perform) { post path, payload.to_json, environment }
 
   before { authorize 'd3m0', 'x' }
+
+  include_examples 'supports siphon'
 
   it 'succeeds' do
     perform
@@ -35,11 +39,11 @@ describe Routemaster::Drain::Mapping do
   end
 
   let(:map) { Routemaster::Dirty::Map.new }
-  
+
   it 'uses dirty map' do
     payload1 = [1,2,3,1].map { |idx| make_event(idx) }.to_json
     payload2 = [1,4,5,2].map { |idx| make_event(idx) }.to_json
-    
+
     post path, payload1, environment
     expect(map.count).to eq(3)
     map.sweep_one("https://example.com/stuff/1") { true }
@@ -48,4 +52,3 @@ describe Routemaster::Drain::Mapping do
     expect(map.count).to eq(4) # 2, 3, 4, 5 (1 is a repeat event, filtered out)
   end
 end
-